Add stopping

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/WinchMotor.java
@@ -144,7 +144,7 @@ public class WinchMotor extends MotorEx {
     }
 
 
-    public void setRawPower(int power) {
+    public void setRawPower(double power) {
         motor.setPower(power);
     }
 }
